Grade/base

+++ src/main/java/shop/gaship/gashipshoppingmall/membergrade/service/MemberGradeServiceImpl.java
@@ -1,59 +1,65 @@
 package shop.gaship.gashipshoppingmall.membergrade.service;
 
 import javax.transaction.Transactional;
-
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import shop.gaship.gashipshoppingmall.member.repository.MemberRepository;
+import shop.gaship.gashipshoppingmall.membergrade.dto.request.MemberGradeAddRequestDto;
+import shop.gaship.gashipshoppingmall.membergrade.dto.request.MemberGradeModifyRequestDto;
+import shop.gaship.gashipshoppingmall.membergrade.dto.response.MemberGradeResponseDto;
+import shop.gaship.gashipshoppingmall.membergrade.dto.response.PageResponseDto;
 import shop.gaship.gashipshoppingmall.membergrade.entity.MemberGrade;
-import shop.gaship.gashipshoppingmall.membergrade.exception.MemberGradeNotFoundException;
+import shop.gaship.gashipshoppingmall.membergrade.exception.*;
 import shop.gaship.gashipshoppingmall.membergrade.repository.MemberGradeRepository;
-import shop.gaship.gashipshoppingmall.membergrade.request.MemberGradeRequest;
 import shop.gaship.gashipshoppingmall.statuscode.entity.StatusCode;
 import shop.gaship.gashipshoppingmall.statuscode.exception.StatusCodeNotFoundException;
 import shop.gaship.gashipshoppingmall.statuscode.repository.StatusCodeRepository;
+import shop.gaship.gashipshoppingmall.statuscode.status.RenewalPeriod;
 
 /**
- * packageName    : shop.gaship.gashipshoppingmall.membergrade.service
- * fileName       : MemberGradeServiceImpl
- * author         : semi
- * date           : 2022/07/09
- * description    :
- * ===========================================================
- * DATE              AUTHOR             NOTE
- * -----------------------------------------------------------
- * 2022/07/09        semi       최초 생성
+ * 회원등급 Service 구현체.
+ *
+ * @author : 김세미
+ * @since 1.0
+ * @see shop.gaship.gashipshoppingmall.membergrade.service.MemberGradeService
  */
 @RequiredArgsConstructor
 @Service
 public class MemberGradeServiceImpl implements MemberGradeService {
     private final MemberGradeRepository memberGradeRepository;
     private final StatusCodeRepository statusCodeRepository;
+    private final MemberRepository memberRepository;
 
     @Transactional
     @Override
-    public void addMemberGrade(MemberGradeRequest request) {
+    public void addMemberGrade(MemberGradeAddRequestDto request) {
         StatusCode renewalPeriod = statusCodeRepository
-                .findById(1)
+                .findByGroupCodeName(RenewalPeriod.GROUP)
                 .orElseThrow(StatusCodeNotFoundException::new);
 
-        MemberGrade memberGrade = MemberGrade.builder()
-                .renewalPeriod(renewalPeriod)
-                .name(request.getName())
-                .accumulateAmount(request.getAccumulateAmount())
-                .build();
+        checkOverlapAccumulateAmount(request.getAccumulateAmount());
 
-        memberGradeRepository.save(memberGrade);
+        if (request.getIsDefault()) {
+            checkExistDefaultMemberGrade();
+            memberGradeRepository.save(MemberGrade.createDefault(renewalPeriod, request));
+        } else {
+            memberGradeRepository.save(MemberGrade.create(renewalPeriod, request));
+        }
     }
 
     @Transactional
     @Override
-    public void modifyMemberGrade(Integer memberGradeNo, MemberGradeRequest request) {
+    public void modifyMemberGrade(MemberGradeModifyRequestDto request) {
         MemberGrade memberGrade = memberGradeRepository
-                .findById(memberGradeNo)
+                .findById(request.getNo())
                 .orElseThrow(MemberGradeNotFoundException::new);
 
-        memberGrade.setName(request.getName());
-        memberGrade.setAccumulateAmount(request.getAccumulateAmount());
+        if (!memberGrade.getAccumulateAmount().equals(request.getAccumulateAmount())) {
+            checkOverlapAccumulateAmount(request.getAccumulateAmount());
+        }
+
+        memberGrade.modifyDetails(request);
 
         memberGradeRepository.save(memberGrade);
     }
@@ -65,6 +71,51 @@ public class MemberGradeServiceImpl implements MemberGradeService {
                 .findById(memberGradeNo)
                 .orElseThrow(MemberGradeNotFoundException::new);
 
+        if (memberGrade.isDefault()) {
+            throw new CannotDeleteDefaultMemberGrade();
+        }
+
+        if (!memberRepository.findByMemberGrades(memberGrade).isEmpty()) {
+            throw new MemberGradeInUseException();
+        }
+
         memberGradeRepository.delete(memberGrade);
+    }
+
+    @Override
+    public MemberGradeResponseDto findMemberGrade(Integer memberGradeNo) {
+        return memberGradeRepository.getMemberGradeBy(memberGradeNo)
+                .orElseThrow(MemberGradeNotFoundException::new);
+    }
+
+    @Override
+    public PageResponseDto<MemberGradeResponseDto> findMemberGrades(Pageable pageable) {
+        return new PageResponseDto<>(memberGradeRepository
+                .getMemberGrades(pageable));
+    }
+
+    /**
+     * 회원등급의 기준누적금액 중복을 방지하기 위해 check 하는 메서드.
+     *
+     * @param accumulateAmount 회원등급 승급의 기준이 되는 기준누적금액 (Long)
+     * @throws AccumulateAmountIsOverlap 기준누적금액을 가지고 있는 회원등급이 이미 존재합니다.
+     * @author 김세미
+     */
+    private void checkOverlapAccumulateAmount(Long accumulateAmount) {
+        if (memberGradeRepository.existsByAccumulateAmountEquals(accumulateAmount)) {
+            throw new AccumulateAmountIsOverlap(accumulateAmount);
+        }
+    }
+
+    /**
+     * 기본회원등급 재등록을 방지하기 위해 check 하는 메서드.
+     *
+     * @throws DefaultMemberGradeIsExist 이미 기본회원등급이 존재합니다.
+     * @author 김세미
+     */
+    private void checkExistDefaultMemberGrade() {
+        if (memberGradeRepository.existsByIsDefaultIsTrue()) {
+            throw new DefaultMemberGradeIsExist();
+        }
     }
 }
